--- a/performance.pl
+++ b/performance.pl
@@ -16,10 +16,8 @@ my $color = "\033[1;32m";
 print "\nBOT STATS:\n";
 my $bot_commands = `docker-compose logs bot | grep processed | wc -l | tr -d '\n'`;
 print "\tNumber of commands received since last deployment - $color$bot_commands\033[0m\n";
-
 my $bot_processing = `docker-compose logs bot | grep processed | sed "s/s//g" | awk '{sum += \$7} END {print sum / NR}' | tr -d '\n'`;
 print "\tAverage time to process a command - $color$bot_processing seconds\033[0m\n";
-
 my $bot_waiting = `docker-compose logs bot | grep retrieved | sed "s/s//g" | awk '{sum += \$7} END {print sum / NR}' | tr -d '\n'`;
 print "\tAverage waiting time for a command - $color$bot_waiting seconds\033[0m\n";
 
@@ -27,18 +25,14 @@ print "\tAverage waiting time for a command - $color$bot_waiting seconds\033[0m\
 print "\nSUBMITTER STATS:\n";
 my $submitter_commands = `docker-compose logs submitter | grep processed | wc -l | tr -d '\n'`;
 print "\tNumber of submissions received since last deployment - $color$submitter_commands\033[0m\n";
-
 my $submitter_processing = `docker-compose logs submitter | grep processed | sed "s/s//g" | awk '{sum += \$7} END {print sum / NR}' | tr -d '\n'`;
 print "\tAverage time to process a submission - $color$submitter_processing seconds\033[0m\n";
-
 my $submitter_waiting = `docker-compose logs submitter | grep retrieved | sed "s/s//g" | awk '{sum += \$7} END {print sum / NR}' | tr -d '\n'`;
 print "\tAverage waiting time for a submission - $color$submitter_waiting seconds\033[0m\n";
 
 # CALCULATOR
 print "\nCALCULATOR STATS:\n";
-
 my $calculator_commands = `docker-compose logs calculator | grep processed | wc -l | tr -d '\n'`;
 print "\tNumber of investments evaluated since last deployment - $color$calculator_commands\033[0m\n";
-
 my $calculator_processing = `docker-compose logs calculator | grep processed | sed "s/s//g" | awk '{sum += \$7} END {print sum / NR}' | tr -d '\n'`;
 print "\tAverage time to process investments - $color$calculator_processing seconds\033[0m\n";