This is a small perl script that analyzes containers' performance. When executed by `./performance.pl`, gives the following output:

```
MemeInvestor script for checking containers' performance
The log analysis will take some time depending on the log size...

BOT STATS:
	Number of commands received since last deployment - 3435
	Average time to process a command - 4.0427 seconds
	Average waiting time for a command - 1448.25 seconds

SUBMITTER STATS:
	Number of submissions received since last deployment - 212
	Average time to process a submission - 7.58873 seconds
	Average waiting time for a submission - 22621.6 seconds

CALCULATOR STATS:
	Number of investments evaluated since last deployment - 1605
	Average time to process investments - 5.29364 seconds
```